// src/server/model/inventoryTxReportGroupLine.ts
/**
 * 
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ITransactionStatus } from './iTransactionStatus';
import { InventoryTxReportLine } from './inventoryTxReportLine';
import { TransactionInfo } from './transactionInfo';

export interface InventoryTxReportGroupLine { 
    txDate?: Date;
    particulars?: string;
    billAmount?: number;
    transaction?: TransactionInfo;
    voucherNo?: string;
    usageReportGroupLines?: Array<InventoryTxReportGroupLine>;
    transactionStatus?: ITransactionStatus;
    details?: string;
    gstin?: string;
    roundAmount?: number;
    inventoryTxReportLines?: Array<InventoryTxReportLine>;
}